Add method `streamControllable(..)`
This method essentially does the same as `stream` previously did, but
instead of returning the stream it does return the SteamControler. This
makes it possible for the abort the streaming on the client side by
calling `close()` on the controller.

The `stream()` uses `streamControllable()` to return the stream, the
method is still keept to not break existing usage of the
method.