Client to server streaming fixes (#99)
* Add method `streamControllable(..)`

This method essentially does the same as `stream` previously did, but
instead of returning the stream it does return the SteamControler. This
makes it possible for the abort the streaming on the client side by
calling `close()` on the controller.

The `stream()` uses `streamControllable()` to return the stream, the
method is still keept to not break existing usage of the
method.

* Fix creating client to server streams

prevents an index out of range exceptions and uses the correct
invocation ids for sent items

* Fix json completion message for client to server streams

This sets either the 'error' or the 'result' field, or neither as it is
specified here:
https://github.com/dotnet/aspnetcore/blob/main/src/SignalR/docs/specs/HubProtocol.md
This prevents exceptions on the server when the client closes the
stream.

* fix build issue

not all changes where committed...